mxc: gpu_viv: Optimize PM routine
We are trying to manage multiple threads that are all trying to
put the GPU in the proper powerstate.  It is possible that while
the hardware is trying to power on the GPU the idle timer will
kick off and first suspend the chip only to wake it back up. This
commit does two things

1)  If the chip is already in the state we are trying to achieve
bail out and do nothing.
2)  If we are powering the chip on or off kill the IDLE timer
right away as we know we don't care about the suspend state because
we are already processing a different state command that is not
suspend.