ENGR00302472-5 video: mxc disp: Let encoder know fbi when enable/disable
This patch adds framebuffer info structure to the mxc display driver
enable/disable callback parameters so that the encoder drivers may
know the framebuffer information.  In this way, the encoder drivers
which may deal with more than one framebuffer are able to figure out
the one it is handling.  A real example is the LDB encoder, which may
deal with two standalone framebuffers at most.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>