mxc: hdmi: DVI mostly working properly
This patches core, CEC, hdmi to properly handle DVI in most
cases.  If in DVI mode we have to turn off hotplug interrupts
or we continually flap on and off.  We should also disable
CEC as it is not supported under DVI, and stop audio if it
is playing.

The one corner case that is still open is if you have HDMI
plugged in and the fb blanks, then while blanked you unplug
your hdmi cable and plug in your dvi monitor.  The display
will come up but will flap until the console is unblanked.