Override function pm_power_off to prevent sending power off signal to PMIC.
The default power off function sends a power off signal to PMIC to cut off all the power supplies. It requires an ON/OFF button to power up the PMIC again until the SNVS power supply goes down. But on curie board, we don't have any ON/OFF button on it. If we don't disable it, we need to wait until the super cap is discharged.

The default power off function could be found in drivers/rtc/rtc-snvs.c