mxc: gpu_viv: Fix poweroff at boot problems
When issuing a poweroff command at boot it is possible for a race
to occur with the interrupt handler (mostly with the VG core) that
results in the poweroff command failing.  Looking at the code there
is already a specific mode to fix this, gcvPOWER_OFF_ATPOWERON.

Why not use this so our unused cores are fully powered off if they
aren't used?