thermal: devfreq_cooling: Allow max_states to be variable
This allows max_state to be set at the time the devfreq cooling
device is instantiated.  This allows events to match the cpufreq
cooling device and gives tighter grained control over the
devices listening.

Currently I am using state 5 as the critical event number.  This
could be dynamic as well but since only the imx_thermal and galcore
use this driver I am okay hardcoding it for testing.

Signed-off-by: Jens Rehsack <sno@netbsd.org>