gpu: galcore: Fix for 2D engine lockups
It seems that if we send a PipeSelect with each command that
doesn't have a Context we can avoid hangups that happening if
you are using intermixed 3d and 2d drawing operations. I believe
it is the flush and stall commands that are actually fixing up
the 2d engine so it will run without issues, but rather than
writing a custom handler I will just re-use the PipeSelect
command.  Technically there is nothing wrong with doing command
submission this way it is just an extra bit of overhead, however
in general benchmarks I have not seen any noticable performance
degredation.