gpu: galcore: implement gcmkONWARNING for error handling
In the previous commit I had set a value to return because it
was not actually an error.  However these errors need to propogate
back through the logic chain so that is not an acceptible fix.

Instead we will implement gcmkONWARNING which if the loglevel is
changed from ERROR to WARNING will report the original error followed
by the rest of the chain marked as a warning.  This allows all the
existing code to go unchanged, but also doesn't report errors to the
logs that are not in fact errors.