gpu: galcore: don't use suspend when idle by default
Now that all the locking is cleaned up and we have removed all
bottlenecks and false returns in the PM it doesn't make much sense
to tear down and bring up the clocks every time the GPU finishes
a command and then another hasn't queued yet.  We still power down
once we hit idle and stay there until the timer runs out which is
really all the power savings we need.