mxc/hdmi-cec: handle HDMI_IH_CEC_STAT0_DONE as CEC_TX_AVAIL
When the hdmi cable is pluged-in in while cec is already set up, we
received a HDMI_IH_CEC_STAT0_DONE event which is unhandled, and will
lock hdmi_cec_write up.

To fix this HDMI_IH_CEC_STAT0_DONE is now handled as CEC_TX_AVAIL
allowing data to be sent.

Signed-off-by: Matthias Rabe <matthias.rabe@sigma-chemnitz.de>